AP-300: refactor image logo provider to separate service
