+++ Components/Adyen/PaymentMethod/ImageLogoProviderInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AdyenPayment\Components\Adyen\PaymentMethod;
+
+interface ImageLogoProviderInterface
+{
+    public function getAdyenImageByType($type): string;
+}
