+++ Components/PaymentMethodService.php
@@ -10,29 +10,11 @@ namespace AdyenPayment\Components;
  */
 class PaymentMethodService
 {
-    const PM_LOGO_FILENAME = [
-        'scheme' => 'card',
-        'yandex_money' => 'yandex'
-    ];
-
     /**
      * PaymentMethodService constructor.
      */
     public function __construct()
     {
-    }
-
-    /**
-     * @param $type
-     * @return string
-     */
-    public function getAdyenImageByType($type)
-    {
-        //Some payment method codes don't match the logo filename
-        if (!empty(self::PM_LOGO_FILENAME[$type])) {
-            $type = self::PM_LOGO_FILENAME[$type];
-        }
-        return sprintf('https://checkoutshopper-live.adyen.com/checkoutshopper/images/logos/%s.svg', $type);
     }
 
     public function getPaymentMethodOptions()
