+++ Resources/services/components.xml
@@ -193,5 +193,9 @@
             <argument type="service" id="AdyenPayment\Doctrine\Writer\PaymentMethodWriter"/>
             <argument type="expression">service('models').getRepository('Shopware\\Models\\Shop\\Shop')</argument>
         </service>
+        <service
+            id="adyen_payment.components.adyen.payment_method.image_logo_provider"
+            class="AdyenPayment\Components\Adyen\PaymentMethod\ImageLogoProvider">
+        </service>
     </services>
 </container>
